Have a single implementation of assigning sources
and also filter them if not set (ie None) so the other
code (validation) does not fail if they are null, so present
but not set (json validator does not like that).
